fix: rename aladin.py into widget.py
this prevents typos like 'from ipyaladin import aladin' instead of 'from ipyaladin import Aladin'
